+++ src/constants.mjs
@@ -6,5 +6,3 @@ export const DOC_NODE_VERSION = process.version;
 // This is the Node.js CHANGELOG to be consumed to generate a list of all major Node.js versions
 export const DOC_NODE_CHANGELOG_URL =
   'https://raw.githubusercontent.com/nodejs/node/HEAD/CHANGELOG.md';
-
-export const DOC_API_LATEST_BASE_URL = 'https://nodejs.org/docs/latest';
