+++ src/generators/llms-txt/template.txt
@@ -1,0 +1,9 @@
+# Node.js Documentation
+
+> Node.js is an open-source, cross-platform JavaScript runtime environment that executes JavaScript code outside a web browser. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient for building scalable network applications.
+
+## Introduction
+__INTRODUCTION__
+
+## API Documentations
+__API_DOCS__
